Projet ISEP/local/requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
Pillow
scipy
torch
torchvision
timm
reportlab
